Convert most to directly use Batteries
The Batteries-Included doesn't provide a way of doing seeks on it's
IO.  In order to use this, the LegacyIO module pulls back in the
necessary IO from the old prelude.  It is still necessary to use
things like Legacy.Buffer.output_buffer since there isn't a way to
shadow bindings in other modules.