Add 'readdir' call
Create a C-stubs library to hold the stubs for the various routines we
will need to call.  Implements the first of these 'readdir' which
returns both the name as well as the inode number of each directory
node.  This will allow the nodes to be sorted by inode number,
allowing for much faster stat performance of large directories.