Create an ordinal index on the file pool
The method #find_index on the file pool will return a small integer
for a given hash.  For a given pool session (without writing), this
integer will uniquely specify a given hash, and the hashes will occupy
the consecutive integers from zero to a specific value.  The hashes
aren't necessarily in any particular order.