Start of seen-database implementation
The make-cache command will traverse a backup, tracking the hashes of
all nodes within each directory.  A compact representation of some
information about the directories will help to avoid needing to
recompute hashes of files that haven't changed.

Currently, the database is just written to a file in /tmp.

This brings in an additional dependency on libsqlite3-ocaml-dev which
is used to store the database.