Use the Seendb cache when backing up
Entries that are unchanged since the last backup can just be stored by
reference instead of recomputing their hashes.

Adds a partially automatic test script.  Given the assumptions are met
(of external git repositories, for example), will run some tests of
the backup process.  The output must be examined visually to determine
if it worked.