Folded expire should be min not max
The cached node data for each directory contains a summarized
expiration time.  This is supposed to indicate the time at which all
of the nodes in that directory can be considered expired.

The correct value is the largest (furthest in the future) expire time
of any node.  Once that value is less than the current time, all nodes
in that directory can be considered expired, and the entire directory
cache dropped.  By incorrectly using the max, the expiration could
potentially discard an entire directory cache as soon as the first
node expires.