Replace 'meter' class with a function interface
The meter's only interface with the logger is a single function to get
the current text.  This works better just as a passed in function.
Move the complex formatting code that is duplicated in each meter into
a helper function, and extract the template code that surrounds the
meter printing.  The new meter code, in addition to be shorter, is
easier to read.