Remove evil hack about Arg.command
The Batteries-Included library includes a helper, Arg.handle, to make
argument processing slightly easier.  Unfortunately, it's type is
hidden, so we can't extract the argument to print out a useful help
message.

Instead of using an evil Obj.magic cast, just make the types what
Arg.parse and Arg.usage want to see, and write the two lines of code
necessary to store the processed arguments into a list.