Use a config file for the default pool
Use the config-file library to manage a config file.  Create a single
option in this for the default pool.  If the config file does not
exist, and the user has permission, the library will write out the
defaults to the config file, as an example of the format.

If the default pool is set to something other than the empty string,
it will used as the default, if no -pool argument is given.