Add -config option for alternate config file
Add a -config option to specify an alternate config file.  To make
this work, we have to parse the global options before reading the
config file.  This then requires that the pool option only override if
it wasn't specified on the command line.  Make the pool a "string
option" itself to match the type from the config file, and reduce the
type confusion over the empty string.