Infrastructure for remote commands
Implements a ping command, and all of the infrastructure around that.
Adds a 'clients' section to the config file which allows the
parameters of remote clients to be specified.  One of these is the
remote command.  The 'sample.conf' contains an example, using ssh to
connect to a local client.  The program is agnostic as to how the
remote client is run.

odump can also be invoked with the 'receiver' argument where it will
expect to be given the remote commands on stdin/stdout.

The protocol implements a single command, `Ping, which is the single
necessary command in this simplistic protocol.  The code manages
protocol versions so that mismatched versions will cause errors rather
than corrupt behavior or crashes.