Get strong entropy for Uuid generation
The Uuidm library relies on [Random.make_self_init] to seed the prng
used to generate Uuids.  As of Ocaml 3.12, this is based entirely on
the time and the pid of the process and parent, which is insufficient
to guarantee global uniqueness.

As an easy way of solving this, we generate a SHA-1 name-based UUID,
using a URL containing a random string.

It would probably be better to generate a direct random UUID, but this
is a little simpler, and doesn't involve understanding the details of
the implementation.

For our purposes, the type fields really don't matter much, anyway.