Fix 'skip' vs 'dup'
The intention was for 'dup' to indicate data that was not backed up
because the sha1 hash was detected in the pool, and 'skip' was data
not backed up because of the seen cache.  Both were being counted as
'skip', and 'dup' was always just zero.  Fix this by incrementing
'dup' when detecting data already present.