Remove all dependency on pxp
The ocaml pxp library depends on camomile, which requires a bunch of
runtime files to be present in the install directory.  This makes
odump more difficult to run without this directory present.  Although
we eliminate support for reading older-style backups, removing the pxp
support does eliminate the need to have these files present.