Pool data file verification
The 'odump verify ...' command will verify specified pool data files
to make sure the integrity of the hashes is correct.

To implement this, add a new method to the chunk_file class that
supports checking a chunk, and returning the next offset.  This
requires some refactoring so that the header information can be
maintained enough to determine the next location.

The verify loop just walks through the given files and makes sure
there are chunks that are correct.