Eliminate (most) warnings
Enable nearly all warnings in the compiler, and eliminate their
occurrences from the source.  The only one remaining is '4': "Fragile
pattern matching: matching that will remain complete even if
additional constructors are added to one of the variant types
matched."  There are a few types in the program with a bunch of
constructors that would be tedious to match fully.