Implement a backup manager
Inspired by the Scala version, this performs all of the steps of a
backup, including snapshots, integrity checks, rsyncs, and the backup
itself to be done in a single command invocation of odump.  The
particulars are described in the config file, which gives the hosts
and the volumes that can be backed up.

The code handles several different filesystem types (with and without
lvm), and tries to cleanup after itself if something goes wrong.