Implement 'newfile' support
This config file option tells odump to start with fresh pool files for
each invocation.  This is useful for certain mirroring situations,
where files should not be appended to once they have been written.

Currently, this has to be set manually in the pool's
metadata/props.txt file with a line such as:

	newfile=true