Fix backup index on index recovery
Add entries to the "metadata/backups.txt" file while performing index
recovery.  This is necessary to truly be able to restore a backup from
just the data files.

To support this, instead of just appending entries to the
"metadata/backups.txt" file, read the existing one, do a set add, and
write the entries out.  This will avoid duplicates that otherwise
could be created when doing the recovery.