Implement SQL pool
The SQL pool is an alternate storage format for the backup pools.
Rather than having to maintain our own index and file formats, store
the bulk of the information in an sqlite3 database.  Blobs that take
more than 100k are written to individual files, as per advice from the
sqlite3 docs.

The pool type is auto-detected upon open, and there are two create
commands, the original for the older file-based pool, and create-sql
which will create a new SQL pool.

The seendb is still kept in separate files, now under poolname.seen
along side the database.