Ignore "test-pool" directory
It is useful to have a 'test-pool' directory.  However, ocamlbuild
likes to descend directories, even symlinks, to look for hygeine
issues.  With a large pool, this can add quite a delay to the initial
build time.

Add a line to the _tags file to tell it to not traverse this directory
at all.